requirements.txt
torch
torchvision
opencv-python
numpy<2
gradio
Pillow
transformers
